refactor(models): split HA tables into ha_glue/ (#342)
Phase 1 Week 1.2 of the platform/ha-glue extraction. Establishes the
first piece of the boundary: SQLAlchemy models for HomeAssistant,
camera, presence, paperless, and radio features move to a new
`ha_glue/models/database.py` while `models/database.py` keeps only
platform-generic tables (conversations, users, knowledge bases,
memory, notifications, reminders, tasks, etc.).

Moved to ha_glue/models/database.py (9 classes + 2 constant blocks):
  CameraEvent, HomeAssistantEntity, Room, RoomDevice, RoomOutputDevice,
  RoomSatellite alias, UserBleDevice, PresenceEvent,
  PaperlessAuditResult, RadioFavorite
  DEVICE_TYPE_* / DEVICE_TYPES / DEFAULT_CAPABILITIES
  OUTPUT_TYPE_* / OUTPUT_TYPES

Stays in models/database.py:
  Conversation, Message, Task, Speaker, SpeakerEmbedding, KnowledgeBase,
  Document, DocumentChunk, ChatUpload, Role, User, KBPermission,
  SystemSetting, IntentCorrection, Notification, NotificationSuppression,
  Reminder, ConversationMemory, EpisodicMemory, KGEntity, KGRelation,
  MemoryHistory

Layering-rule fixes
-------------------
Removed two platform -> ha-glue dependencies that violated the boundary:

1. Notification.room_id had `ForeignKey("rooms.id")` and a
   `relationship("Room")`. The relationship was declared but never
   read by application code. Dropped both — the column is now a plain
   Integer. Runtime lookup via the hook system happens in ha_glue.

2. Reminder.room_id had the same pattern. Same fix.

Neither change affects behavior because neither `.room` accessor was
used anywhere; only `room_id` and `room_name` columns are read and
written. The FK was also never enforced by ORM cascade.

Backwards compatibility
-----------------------
`models/database.py` ends with a try/except compat re-export block
that pulls the HA classes back in from `ha_glue.models.database`, so
existing consumer code (`from models.database import Room` in 11
files across services/ and api/routes/) keeps working during the
Week 2-4 transition. The re-exports will be removed in Week 4 once
every consumer has been migrated to import directly from ha_glue and
the CI lint rule forbids platform -> ha-glue imports.

Verification
------------
- AST parse: both files parse clean
- Import smoke test: 10 of 11 consumer files import successfully
  (the 11th fails only on missing `redis` in the local dev env, a
  pre-existing issue)
- `configure_mappers()`: succeeds, 31 tables registered against
  Base.metadata (was 31 before, identity preserved)
- `tests/backend/test_models.py`: 2 passed, 21 errors — identical to
  pristine `main`, all errors are pre-existing SQLite/TSVECTOR
  incompatibility in the test fixture (unrelated to this change)
- Re-exports verified: `from models.database import Room` and
  `from ha_glue.models.database import Room` return the same class
  object

Layering rule status
--------------------
After this commit, the only remaining platform -> ha_glue violations
are two local (function-scope) imports in `services/reminder_service.py`
and `api/lifecycle.py` that do `from models.database import Room` to
resolve room names. These still work via the compat re-export, and
will be extracted via hooks in the Week 1.2 agent-router / lifecycle
splits that land in follow-up commits.

Refs
----
- Tracking: ebongard/renfield#342 (parent Phase 1 issue)
- Parent plan: X-idra-Systems-GmbH/reva#133 (Renfield OSS launch)
- Boundary doc: X-idra-Systems-GmbH/reva docs/architecture/renfield-platform-boundary.md
